Add files via upload
PLEASE FIND THE BUG!!

When I do the loop on the urls it raises an error. The code works, because if I execute it with all the different urls (without executing any of them before) it does everything! The error rises when I try to run them consequently, I think there is a bug in one function where maybe a variable is recalled so that doesn't change every time!